rename profile page to settings page and all related components
