+++ src/app-ui/pages/settings/subpages/TermsOfServicePage.tsx
